requirements.txt
Flask>=2.0.1
numpy>=1.21.0
pandas>=1.3.0
pytest>=6.2.5
python-dotenv>=0.19.0 